Added performance gain for getting table columns by using one single call to the database with INFORMATION_SCHEMA Separated "bigint" to Int64 and "bigint unsigned" to UInt64
