+++ Src/Simple.Data.Mysql.Test/Simple.Data.Mysql.Test.csproj
@@ -107,7 +107,6 @@
     <Compile Include="MysqlConnectorHelperTest.cs" />
     <Compile Include="OrderDetailTests.cs" />
     <Compile Include="SchemaDataProviderTests\40SchemaDataProviderTests.cs" />
-    <Compile Include="SchemaDataProviderTests\50SchemaDataProviderTests.cs" />
     <Compile Include="SchemaDataProviderTests\SchemaDataProviderSelectionTest.cs" />
     <Compile Include="SchemaTests\ColumnInfoCreatorTest.cs" />
     <Compile Include="SchemaTests\DatabaseSchemaTests.cs" />
