+++ Src/Simple.Data.Mysql/MysqlColumnInfo.cs
@@ -21,7 +21,8 @@ namespace Simple.Data.Mysql
                                                                             {"mediumint unsigned", DbType.UInt32},
                                                                             {"int", DbType.Int32},
                                                                             {"int unsigned", DbType.UInt32},
-                                                                            {"bigint", DbType.UInt64},
+                                                                            {"bigint", DbType.Int64},
+                                                                            {"bigint unsigned", DbType.UInt64},
                                                                             {"float", DbType.Single},
                                                                             {"double", DbType.Double},
                                                                             {"decimal", DbType.Decimal},
