+++ Src/Simple.Data.Mysql/Properties/AssemblyInfo.cs
@@ -8,6 +8,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTitle("Simple.Data.Mysql")]
 [assembly: AssemblyDescription("")]
 [assembly: InternalsVisibleTo("Simple.Data.Mysql.Test")]
+[assembly: InternalsVisibleTo("Simple.Data.Mysql5.Test.Features")]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("eb41e7d5-70b2-4d9c-bf03-5266688d3701")]
