+++ Src/Simple.Data.Mysql5.Test.Features/packages.config
@@ -1,5 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Castle.Core" version="3.0.0.4001" targetFramework="net40" />
+  <package id="FakeItEasy" version="1.7.4507.61" targetFramework="net40" />
+  <package id="FluentAssertions" version="1.7.1.1" targetFramework="net40" />
   <package id="NUnit" version="2.5.9.10348" targetFramework="net40" />
   <package id="Simple.Data.Ado" version="0.19.0.0" targetFramework="net40" />
   <package id="Simple.Data.Core" version="0.19.0.0" targetFramework="net40" />
